+++ src/main/java/model/Constant.java
@@ -1,3 +1,5 @@
+package model;
+
 public class Constant {
 
     public static final int LOTTO_PRICE = 1000;
